+++ src/pages/[...locale]/feed.xml.ts
@@ -3,6 +3,7 @@ import { experimental_AstroContainer as AstroContainer } from "astro/container";
 import { i18n } from "astro:config/client";
 import { getCollection, render } from "astro:content";
 import { getRelativeLocaleUrl } from "astro:i18n";
+import { monolocale } from "astro:locales";
 import { Feed } from "feed";
 import config from "$config";
 import i18nit from "$i18n";
@@ -39,39 +40,42 @@ export const GET: APIRoute = async ({ site, params }) => {
 	// Aggregate items from specified sections
 	let items = [];
 
-	if (config.feed?.section?.includes("note") || config.feed?.section === "*" || config.feed?.section === undefined) {
+	// Determine which sections to include
+	const sections = config.feed?.section || "*";
+
+	if (sections === "*" || sections.includes("note")) {
 		const notes = await getCollection("note", note => {
-			// Extract language from the file path structure
-			const [locale, ...id] = note.id.split("/");
-
-			// Attach locale and link
-			(<any>note).link = new URL(getRelativeLocaleUrl(locale, `/note/${id.join("/")}`), site).toString();
-
 			// Apply filtering criteria
 			const published = !note.data.draft; // Exclude draft posts
-			const localed = language === locale; // Language filter
+			const localed = monolocale || note.id.split("/")[0] === language; // Language filter
 
 			// Include note only if it passes all filters
 			return published && localed;
 		});
 
+		// Attach locale and link for each note
+		notes.forEach(note => {
+			const id = monolocale ? note.id : note.id.split("/").slice(1).join("/");
+			Reflect.set(note, "link", new URL(getRelativeLocaleUrl(language, `/note/${id}`), site).toString());
+		});
+
 		items.push(...notes);
 	}
 
-	if (config.feed?.section?.includes("jotting") || config.feed?.section === "*" || config.feed?.section === undefined) {
+	if (sections === "*" || sections.includes("jotting")) {
 		const jottings = await getCollection("jotting", jotting => {
-			// Extract language from the file path structure
-			const [locale, ...id] = jotting.id.split("/");
-
-			// Attach locale and link
-			(<any>jotting).link = new URL(getRelativeLocaleUrl(locale, `/jotting/${id.join("/")}`), site).toString();
-
 			// Apply filtering criteria
 			const published = !jotting.data.draft; // Exclude draft posts
-			const localed = language === locale; // Language filter
+			const localed = monolocale || jotting.id.split("/")[0] === language; // Language filter
 
-			// Include note only if it passes all filters
+			// Include jotting only if it passes all filters
 			return published && localed;
+		});
+
+		// Attach locale and link for each jotting
+		jottings.forEach(jotting => {
+			const id = monolocale ? jotting.id : jotting.id.split("/").slice(1).join("/");
+			Reflect.set(jotting, "link", new URL(getRelativeLocaleUrl(language, `/jotting/${id}`), site).toString());
 		});
 
 		items.push(...jottings);